requirements.txt
python-telegram-bot==20.7
langchain==0.3.14
langchain-core==0.3.29
langgraph==0.2.60
langgraph-checkpoint==2.0.15
langgraph-checkpoint-sqlite==2.0.7
langgraph-prebuilt==0.1.1
langchain-anthropic==0.3.1
langchain-openai==0.3.0
playwright==1.47.0
aiohttp==3.10.5
aiosqlite==0.20.0
python-dotenv==1.0.1
pillow==10.4.0
httpx==0.25.2
pydantic==2.8.2